FIXED: race condition in messagebuffers leading to fewer buffers reaching full during test.  fixed with seq_cst memory order throughout (IMPORTANT), ensure that all branches in append have atomic_thread_fence, updaterCount updated before fence, block as early as possible, check for blocked buffer as early as possible.  In MessageBuffers, use an array of atomic bool to control when a thread can access a Buffer during Buffer Swap - threads will spinlock on BufferReady until swap is done (likely Important), on release check to make sure all updaters are done (via spinlock on updaterCount.  IMPORTANT), also get a pointer to Buffer at beginning of append, instead of accessing the array which may change during the invocation of append. Buffer swap-in still requires mutex lock.  Changed to seq_cst memory order in threadsafe_queue, comm layer, buffer, message_buffers, and partitioner.  WIP: identified that Buffer contains 1 less than expected entries - could be causing the problem where number of buffer is 1 short as well (try with 25 entries, 1000 iterations)

+++ src/io/test/check_buffer.cpp
@@ -31,7 +31,7 @@ int main(int argc, char** argv) {
 
   printf("TEST insert\n");
   int start = -1, end = -1;
-  int i;
+  unsigned int i;
   for (i = 0; i < 9000; ++i) {
     val = static_cast<valType>(i);
     if (! (tlBuffer.append(&val, sizeof(valType)).first) ) {
@@ -41,7 +41,7 @@ int main(int argc, char** argv) {
   end = i;
   if (start != 8192/sizeof(valType)) printf("\tFAILED can't insert (full) from %d to %d!\n", start, end);
   const valType* temp = reinterpret_cast<const valType*>(tlBuffer.getData());
-  for (int j = 0; j < (8192/sizeof(valType)); ++j) {
+  for (int j = 0; j < int(8192/sizeof(valType)); ++j) {
     if (temp[j] != j) {
       printf("ERROR: result is not expected.  expected: %d actual %d\n", j, temp[j]);
       //assert(false);
@@ -69,7 +69,7 @@ int main(int argc, char** argv) {
 
 
   temp = reinterpret_cast<const valType*>(tlBuffer.getData());
-  for (int j = 0; j < (8192/sizeof(valType)); ++j) {
+  for (int j = 0; j < int(8192/sizeof(valType)); ++j) {
     if (temp[j] != j) {
       //printf("ERROR: result is not expected.  expected: %d actual %d\n", j, temp[j]);
       assert(false);
@@ -192,7 +192,7 @@ int main(int argc, char** argv) {
 
 
   temp = reinterpret_cast<const valType*>(tlBuffer.getData());
-  for (int j = 0; j < (8192/sizeof(valType)); ++j) {
+  for (int j = 0; j < int(8192/sizeof(valType)); ++j) {
     assert(temp[j] == j);
   }
   printf("CONTENT Passes check after move\n");
